--- app/api/genre-api/src/main/java/com/example/config/GenreApiConfig.java
@@ -1,10 +1,0 @@
-package com.example.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(basePackages = "com.example")
-public class GenreApiConfig {
-
-}
