+++ app/api/show-api/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    implementation project(':app:api:common-api')
+}
