+++ app/api/show-api/src/main/java/com/example/config/ShowApiConfig.java
@@ -5,6 +5,6 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan(basePackages = "com.example")
-public class ArtistApiConfig {
+public class ShowApiConfig {
 
 }
