+++ app/api/show-api/src/main/java/com/example/controller/artist/ArtistController.java
@@ -1,7 +1,7 @@
-package com.example.controller;
+package com.example.controller.artist;
 
-import com.example.controller.dto.request.GenreSubscriptionApiRequest;
-import com.example.controller.dto.request.GenreUnsubscriptionApiRequest;
+import com.example.controller.artist.dto.request.ArtistSubscriptionApiRequest;
+import com.example.controller.artist.dto.request.ArtistUnsubscriptionApiRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -14,14 +14,14 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/genres")
-@Tag(name = "장르")
-public class GenreController {
+@RequestMapping("/api/v1/artists")
+@Tag(name = "아티스트")
+public class ArtistController {
 
     @PostMapping("/subscribe")
     @Operation(summary = "구독하기")
-    public ResponseEntity<Void> subscribe(
-        @Valid @RequestBody GenreSubscriptionApiRequest request
+    public ResponseEntity<Void> bulkSubscribe(
+        @Valid @RequestBody ArtistSubscriptionApiRequest request
     ) {
         return ResponseEntity.noContent().build();
     }
@@ -29,7 +29,7 @@ public class GenreController {
     @PostMapping("/unsubscribe")
     @Operation(summary = "구독 취소하기")
     public ResponseEntity<Void> unsubscribe(
-        @Valid @RequestBody GenreUnsubscriptionApiRequest request
+        @Valid @RequestBody ArtistUnsubscriptionApiRequest request
     ) {
         return ResponseEntity.noContent().build();
     }
