+++ app/api/src/main/java/org/example/config/ApiConfig.java
@@ -1,11 +1,11 @@
 package org.example.config;
 
-import com.example.config.ArtistApiConfig;
+import com.example.config.ShowApiConfig;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 @Configuration
-@Import({UserApiConfig.class, ArtistApiConfig.class})
+@Import({UserApiConfig.class, ShowApiConfig.class})
 public class ApiConfig {
 
 }
