+++ settings.gradle
@@ -12,8 +12,7 @@ include (":app:domain:show-domain")
 include(":app:api")
 include (":app:api:common-api")
 include(":app:api:user-api")
-include(":app:api:artist-api")
-include(":app:api:genre-api")
+include(":app:api:show-api")
 
 include (":app:infrastructure")
 include (":app:infrastructure:redis")
